fix custom prompt example
As it stands the current custom prompt example doesn't set an initial value because the constructor is looking for a value called `sprays`. Another possible fix is to rename the `initial` prompt to `sprays` while setting up the prompt; however, it seems better to me to make custom prompts that match conventions set by the default set of prompts.