Fix .validate promise return type
When using .validate with promises/async, the return type needs to allow the function to return boolean or string, rather than just boolean or just string. Otherwise, the string option isn't useful (since you can return a string as an error, but can't return `true` to indicate a valid prompt).